requirements.txt
# FastAPI and server dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database dependencies
sqlalchemy==2.0.23
alembic==1.12.1
# Use psycopg2-binary for Windows compatibility
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Supabase integration
supabase==2.0.2

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Data validation and serialization
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP client for external APIs
httpx==0.25.2

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development and code quality
black==23.11.0
flake8==6.1.0
mypy==1.7.1
isort==5.12.0

# Logging and monitoring
loguru==0.7.2

# CORS support
fastapi-cors==0.0.6
